% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLS-pkg.r
\docType{package}
\name{OLS-pkg}
\alias{OLS-pkg}
\alias{OLS-pkg-package}
\title{package OLSfit}
\description{
Dans ce package, on trouve les fonctions trucs bidule et muche. Elles font ceci et cela. Les classes d'objets Robert et MItchoum font etc.
}
