% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLS.r
\name{OLS}
\alias{OLS}
\alias{OLS}
\title{Description of the function}
\usage{
OLS(X, y)

OLS(X, y)
}
\arguments{
\item{X}{A matrix}

\item{y}{A vector}

\item{a}{paramter}
}
\value{
what is returned

a vector of coefficients
}
\description{
Description of the function

Compute the OLS esstimator
}
\examples{
n <- 100; p <- 10
X <- matrix(rnorm(n*p), n, p)
beta <- runif(p, 1, 2) * sample(c(-1,1), p, rep=TRUE)
y <- X \%*\% beta + rnorm(n)
OLS(X,y)
}
\author{
SB
}
