% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{OLScpp}
\alias{OLScpp}
\title{Compute the OLS esstimator}
\usage{
OLScpp(X, y)
}
\arguments{
\item{X}{A matrix}

\item{y}{A vector}
}
\value{
a vector of coefficients
}
\description{
Compute the OLS esstimator
}
\examples{
n <- 100; p <- 10
X <- matrix(rnorm(n*p), n, p)
beta <- runif(p, 1, 2) * sample(c(-1,1), p, rep=TRUE)
y <- X \%*\% beta + rnorm(n)
OLSfast(X,y)
}
